refactor(tsconfig.json): mapped references to containers to hoc
