+++ src/pages/lobby.tsx
@@ -1,5 +1,5 @@
 import { Heading, Flex, Progress } from '@chakra-ui/react';
-import withAuth from '@/containers/withAuthentication';
+import withAuth from 'src/hoc/withAuthentication';
 import { useRouter } from 'next/router';
 import { io } from 'socket.io-client';
 import { useState } from 'react';
