+++ src/pages/postGame.tsx
@@ -1,5 +1,5 @@
 import { Heading, Spinner } from '@chakra-ui/react';
-import withAuth from '@/containers/withAuthentication';
+import withAuth from 'src/hoc/withAuthentication';
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 import {
